[#1] 변경된 구조에 따라 apiClient kafka-consumer로 이동
