--- api-server/src/main/java/com/pinnochio/santaclothes/apiserver/apiclient/fileserver/dto/FileServerRequest.java
@@ -1,12 +1,0 @@
-package com.pinnochio.santaclothes.apiserver.apiclient.fileserver.dto;
-
-import org.springframework.core.io.Resource;
-
-import lombok.Value;
-
-@Value
-public class FileServerRequest {
-	Long userId;
-
-	Resource file;
-}
