--- api-server/src/main/java/com/pinnochio/santaclothes/apiserver/apiclient/fileserver/type/FileServerStatus.java
@@ -1,6 +1,0 @@
-package com.pinnochio.santaclothes.apiserver.apiclient.fileserver.type;
-
-public enum FileServerStatus {
-	SUCCEED,
-	FAILED,
-}
