--- api-server/src/main/java/com/pinnochio/santaclothes/apiserver/config/KafkaConfig.java
@@ -1,10 +1,0 @@
-package com.pinnochio.santaclothes.apiserver.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-
-@Configuration
-@EnableKafka
-public class KafkaConfig {
-
-}
