--- api-server/src/main/java/com/pinnochio/santaclothes/apiserver/service/FileUploadService.java
@@ -1,13 +1,0 @@
-package com.pinnochio.santaclothes.apiserver.service;
-
-import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.client.WebClient;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class FileUploadService {
-
-	private final WebClient webClient;
-}
