+++ api-server/src/main/java/com/pinocchio/santaclothes/apiserver/config/KafkaConfig.java
@@ -1,0 +1,10 @@
+package com.pinocchio.santaclothes.apiserver.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.kafka.annotation.EnableKafka;
+
+@Configuration
+@EnableKafka
+public class KafkaConfig {
+
+}
