--- api-server/src/main/java/com/pinocchio/santaclothes/apiserver/type/ApiResponseStatus.java
@@ -1,6 +1,0 @@
-package com.pinocchio.santaclothes.apiserver.type;
-
-public enum ApiResponseStatus {
-	SUCCESS,
-	FAILED,
-}
