--- api-server/src/test/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.env.EnvironmentPostProcessor=com.pinnochio.santaclothes.apiserver.test.WireMockPostProcessor
