--- kafka-consumer/src/main/java/com/pinnochio/santaclothes/ConsumerApplication.java
@@ -1,13 +1,0 @@
-package com.pinnochio.santaclothes;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ConsumerApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ConsumerApplication.class, args);
-	}
-
-}
