+++ kafka-consumer/src/main/java/com/pinocchio/santaclothes/consumer/ConsumerApplication.java
@@ -1,0 +1,13 @@
+package com.pinocchio.santaclothes.consumer;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ConsumerApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ConsumerApplication.class, args);
+	}
+
+}
