--- kafka-consumer/src/test/java/com/pinnochio/santaclothes/ConsumerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.pinnochio.santaclothes;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ConsumerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
