+++ kafka-consumer/src/test/java/com/pinocchio/santaclothes/consumer/ConsumerApplicationTests.java
@@ -1,0 +1,13 @@
+package com.pinocchio.santaclothes.consumer;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ConsumerApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
