+++ kafka-consumer/src/test/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.env.EnvironmentPostProcessor=com.pinocchio.santaclothes.consumer.test.WireMockPostProcessor
