+++ kafka-consumer/src/test/resources/application-test.yml
@@ -1,0 +1,11 @@
+spirng:
+  profiles:
+    active: test
+
+api:
+  file-server:
+    url: http://${host.file-server.api}
+
+host:
+  file-server:
+    api: localhost:${wiremock.server.port}
